General improvements and defaults
Refactored the logic for processing activities to return the database connection back to the pool once it is no longer required.

Previously the connection was held onto until all the activities had been processed.

Adjusted the default afterDate for activities to be within the last 3 days.  This change will prevent issues where users with little or no Strava history could launch API calls spanning months or even years of prior Fitbit activities.

This change is inline with the goal of having this application focus solely on handling "recent" activities.

Perhaps in a future version support for processing a more thorough backlog of activities could be added.  But at the moment that is beyond the scope of this application.