Update Strava date parameters
The using start_date can lead to situations where UTC time is used.  Creating unexpected patterns.

Therefore we instead use local time as that SHOULD agree with the time stamps for Fitbit provided that accounts are correctly setup with agreeing local timezones.

A future update could change to converting all times to UTC for a more standardized comparison.  However it is the believable that this pattern will better match the intended flow.